Update Q3Tests .sh, .txt, .java.list, sh,list
Q3Tests.java.list is the .java source, annotated with which lines in
Q3Tests.sh test which lines in the .java source. its a Unit test or
white box scorecard. Doing this review suggested one set of additional
tests- I hadn’t set out to make it that way, but the code, as it
stands, will process sequences of multiple category indicators and
prices. So I’ve added a test for that ability, and that concludes our
unit tests for the evening,